used tailwind flexboxes, simplified CSS, broke CaseSummary into React components
